refactor: move ObjectOrArray type to utils dir
