+++ src/types/common.ts
@@ -9,12 +9,6 @@
 
 */
 
-/**
- *
- */
-// export type ObjectOrArray<T> = T[] | Record<string, T | Record<string, T> | T[]>
-export type ObjectOrArray<T> = T[] | Record<string, T>
-
 /*
  * Basic theme object shape
  *
